vkd3d: Fix misc UCRT warnings.
Found while testing compilation with MinGW UCRT.

+++ tests/d3d12_device.c
@@ -2741,18 +2741,18 @@ void test_misc_agility_sdk_feature_checks(void)
 
     hr = ID3D12Device_CheckFeatureSupport(device, D3D12_FEATURE_APPLICATION_SPECIFIC_DRIVER_STATE, &driver_state, sizeof(driver_state));
     /* Check that we recognize the query. Support doesn't matter. */
-    ok(SUCCEEDED(hr), "Got hr #%x\n", (unsigned int)hr);
+    ok(SUCCEEDED(hr), "Got hr #%x\n", (int)hr);
 
     hr = ID3D12Device_CheckFeatureSupport(device, D3D12_FEATURE_BYTECODE_BYPASS_HASH_SUPPORTED, &bytecode_bypass, sizeof(bytecode_bypass));
     /* Check that we recognize the query. Support doesn't matter. */
-    ok(SUCCEEDED(hr), "Got hr #%x\n", (unsigned int)hr);
+    ok(SUCCEEDED(hr), "Got hr #%x\n", (int)hr);
 
     /* Somewhat a special case, returns E_FAIL on failure. */
     hr = ID3D12Device_CheckFeatureSupport(device, D3D12_FEATURE_SHADER_CACHE_ABI_SUPPORT, &shadercache, sizeof(shadercache));
-    ok(hr == S_OK || hr == E_FAIL, "got hr #%x\n", hr);
+    ok(hr == S_OK || hr == E_FAIL, "got hr #%x\n", (int)hr);
 
     hr = ID3D12Device_CheckFeatureSupport(device, D3D12_FEATURE_D3D12_OPTIONS22, &options22, sizeof(options22));
-    ok(hr == S_OK, "Unexpected hr #%x\n", hr);
+    ok(hr == S_OK, "Unexpected hr #%x\n", (int)hr);
     ok(options22.Max1DDispatchSize >= 0xffff, "Unexpected dispatch size: %u\n", options22.Max1DDispatchSize);
 
     if (SUCCEEDED(ID3D12Device_CheckFeatureSupport(device, D3D12_FEATURE_D3D12_OPTIONS7, &options7, sizeof(options7))) &&
